requirements_minimal.txt
# Core dependencies - Essential for API
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Data processing - Python 3.13 compatible
pandas==2.2.3
numpy==2.1.3
pillow==11.0.0
scipy==1.14.1

# Machine Learning - Essential
tensorflow==2.20.0
scikit-learn==1.5.2
opencv-python==4.10.0.84
matplotlib==3.9.2
seaborn==0.13.2

# HTTP client
httpx==0.27.2
aiofiles==24.1.0